requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0

# MCP (Model Context Protocol)
fastmcp>=0.2.0
mcp>=1.0.0

# FastAPI and web server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# UI framework
streamlit>=1.40.0

# HTTP clients
requests>=2.31.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0

# OpenAI
openai>=1.3.0